fix: extract envHelper unit tests and delete redundant refactor test file (#377)
Closes #377
